requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-cors-headers==3.1.1
django-environ==0.4.5
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
